Split browse media
